requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
requests>=2.31.0
python-dotenv>=1.0.0
label-studio-sdk>=0.0.34
dj-database-url>=2.2.0
whitenoise>=6.8.2
gunicorn>=23.0.0
